Implemented CorrelatorStack, bugfixes
There is still an open issue on what is the correct calculation
of the remaining moment depending on the setting of imax.

Currently solve_all (now moved to CorrelatorStack) contains a lot
of debugging/assert code that will eventually be removed. The code
already prepares for save/restore functionality to be added later.