Two important bug-fixes for tagged particle!
This fixes the calculation of the tagged-particle memory
kernel, there was a sign error in a term.

This fixes the calculation of MSD and NGP, since the base-model
memory kernel in the solver was not initialized correctly.

Now the results for the tagged-particle correlator and the NGP
are in agreement with the results published by Mayr et al.
up to perhaps minor numerical differences that have not been checked.