Identified base.phi bug.
This commit does not fix the bug, it will need some restructuring.

The code in f12.py fails the second solution using stored phi(t)
when the nonergodicity parameter has been called before. The
set_base() call in the nonergodicity_parameter initialization is
the culprit. For now, just recognize that the re-using of models
like this does not work.

A fix is probably to rethink the signatures of the jit-compiled
functions to never reference global variables.