Test: 2d model with pre-hashed x values.
In the 2d model, for each q and k, we need to figure out
the x values over which to integrate, and evaluate arcsin
and similar functions on those x; currently this is calculated
on the fly repeatedly. This commit here contains a version
that implements pre-hashing the x values, but apparently this
version is even a bit slower because (I guess) performance is
lost in the many memory lookups. The implementation will go
away with the next commit, and I don't even bother creating a
separate branch, because the code really becomes a bit uglier
as well with all these hashes.